// src/types/stable_horde.d.ts
  export interface UserDetails {
    /** The user's unique Username. It is a combination of their chosen alias plus their ID. */
    username?: string;
    /** The user unique ID. It is always an integer. */
    id?: number;
    /** The amount of Kudos this user has. The amount of Kudos determines the priority when requesting image generations. */
    kudos?: number;
    /** (Privileged) The amount of Evaluating Kudos this untrusted user has from generations and uptime. When this number reaches a prespecified threshold, they automatically become trusted. */
    evaluating_kudos?: number;
    /** How many concurrent generations this user may request. */
    concurrency?: number;
    /** Whether this user has been invited to join a worker to the horde and how many of them. When 0, this user cannot add (new) workers to the horde. */
    worker_invited?: number;
    /**
     * This user is a Horde moderator.
     * @example false
     */
    moderator?: boolean;
    kudos_details?: UserKudosDetails;
    /** How many workers this user has created (active or inactive) */
    worker_count?: number;
    worker_ids?: string[];
    monthly_kudos?: MonthlyKudos;
    /**
     * This user is a trusted member of the Horde.
     * @example false
     */
    trusted?: boolean;
    /**
     * This user has been flagged for suspicious activity.
     * @example false
     */
    flagged?: boolean;
    /**
     * (Privileged) How much suspicion this user has accumulated
     * @example 0
     */
    suspicious?: number;
    /**
     * If true, this user has not registered using an oauth service.
     * @example false
     */
    pseudonymous?: boolean;
    /**
     * (Privileged) Contact details for the horde admins to reach the user in case of emergency.
     * @example email@example.com
     */
    contact?: string;
    /**
     * How many seconds since this account was created
     * @example 60
     */
    account_age?: number;
    usage?: UsageDetails;
    contributions?: ContributionsDetails;
    records?: UserRecords;
  }
  
  export interface UserKudosDetails {
    /** The ammount of Kudos accumulated or used for generating images. */
    accumulated?: number;
    /** The amount of Kudos this user has given to other users. */
    gifted?: number;
    /** The amount of Kudos this user has been given by the Horde admins. */
    admin?: number;
    /** The amount of Kudos this user has been given by other users. */
    received?: number;
    /** The amount of Kudos this user has received from recurring rewards. */
    recurring?: number;
    /** The amount of Kudos this user has been awarded from things like rating images. */
    awarded?: number;
  }
  
  export interface MonthlyKudos {
    /** How much recurring Kudos this user receives monthly. */
    amount?: number;
    /**
     * Last date this user received monthly Kudos.
     * @format date-time
     */
    last_received?: string;
  }
  
  export interface UsageDetails {
    /** How many megapixelsteps this user has requested */
    megapixelsteps?: number;
    /** How many images this user has requested */
    requests?: number;
  }
  
  export interface ContributionsDetails {
    /** How many megapixelsteps this user has generated */
    megapixelsteps?: number;
    /** How many images this user has generated */
    fulfillments?: number;
  }
  
  export interface UserRecords {
    usage?: UserThingRecords;
    contribution?: UserThingRecords;
    fulfillment?: UserAmountRecords;
    request?: UserAmountRecords;
  }
  
  export interface UserThingRecords {
    /** How many megapixelsteps this user has generated or requested */
    megapixelsteps?: number;
    /** How many token this user has generated or requested */
    tokens?: number;
  }
  
  export interface UserAmountRecords {
    /** How many images this user has generated or requested */
    image?: number;
    /** How many texts this user has generated or requested */
    text?: number;
    /** How many texts this user has generated or requested */
    interrogation?: number;
  }
// ...
  export type WorkerDetails = WorkerDetailsLite & {
    /** How many images this worker has generated. */
    requests_fulfilled?: number;
    /** How many Kudos this worker has been rewarded in total. */
    kudos_rewards?: number;
    kudos_details?: WorkerKudosDetails;
    /** The average performance of this worker in human readable form. */
    performance?: string;
    /** How many threads this worker is running. */
    threads?: number;
    /** The amount of seconds this worker has been online for this Horde. */
    uptime?: number;
    /**
     * When True, this worker will not pick up any new requests
     * @example false
     */
    maintenance_mode?: boolean;
    /**
     * (Privileged) When True, this worker not be given any new requests.
     * @example false
     */
    paused?: boolean;
    /**
     * Extra information or comments about this worker provided by its owner.
     * @example https://dbzer0.com
     */
    info?: string;
    /** Whether this worker can generate NSFW requests or not. */
    nsfw?: boolean;
    /**
     * Privileged or public if the owner has allowed it. The alias of the owner of this worker.
     * @example username#1
     */
    owner?: string;
    /** The worker is trusted to return valid generations. */
    trusted?: boolean;
    /** The worker's owner has been flagged for suspicious activity. This worker will not be given any jobs to process. */
    flagged?: boolean;
    /**
     * (Privileged) How much suspicion this worker has accumulated
     * @example 0
     */
    suspicious?: number;
    /**
     * How many jobs this worker has left uncompleted after it started them.
     * @example 0
     */
    uncompleted_jobs?: number;
    models?: string[];
    team?: TeamDetailsLite;
    /**
     * (Privileged) Contact details for the horde admins to reach the owner of this worker in emergencies.
     * @example email@example.com
     */
    contact?: string;
    /**
     * The bridge agent name, version and website
     * @example AI Horde Worker:11:https://github.com/db0/AI-Horde-Worker
     */
    bridge_agent: string;
    /**
     * The maximum pixels in resolution this worker can generate
     * @example 262144
     */
    max_pixels?: number;
    /** How many megapixelsteps this worker has generated until now */
    megapixelsteps_generated?: number;
    /** If True, this worker supports and allows img2img requests. */
    img2img?: boolean;
    /** If True, this worker supports and allows inpainting requests. */
    painting?: boolean;
    /** If True, this worker supports and allows post-processing requests. */
    "post-processing"?: boolean;
    /**
     * The maximum tokens this worker can generate
     * @example 80
     */
    max_length?: number;
    /**
     * The maximum tokens this worker can read
     * @example 80
     */
    max_context_length?: number;
    /** How many tokens this worker has generated until now */
    tokens_generated?: number;
  };
  
  export interface WorkerDetailsLite {
    /**
     * The Type of worker this is.
     * @example image
     */
    type?: "image" | "text" | "interrogation";
    /** The Name given to this worker. */
    name?: string;
    /** The UUID of this worker. */
    id?: string;
    /** True if the worker has checked-in the past 5 minutes. */
    online?: boolean;
  }
  
  export interface WorkerKudosDetails {
    /** How much Kudos this worker has received for generating images */
    generated?: number;
    /** How much Kudos this worker has received for staying online longer */
    uptime?: number;
  }
  
  export interface TeamDetailsLite {
    /** The Name given to this team. */
    name?: string;
    /** The UUID of this team. */
    id?: string;
  }
  
  export interface ModifyWorkerInput {
    /** Set to true to put this worker into maintenance. */
    maintenance?: boolean;
    /** if maintenance is True, you can optionally provide a message to be used instead of the default maintenance message, so that the owner is informed. */
    maintenance_msg?: string;
    /** (Mods only) Set to true to pause this worker. */
    paused?: boolean;
    /** You can optionally provide a server note which will be seen in the server details. No profanity allowed! */
    info?: string;
    /** When this is set, it will change the worker's name. No profanity allowed! */
    name?: string;
    /**
     * The team towards which this worker contributes kudos.  It an empty string ('') is passed, it will leave the worker without a team. No profanity allowed!
     * @example 0bed257b-e57c-4327-ac64-40cdfb1ac5e6
     */
    team?: string;
  }
  
  export interface ModifyWorker {
    /** The new state of the 'maintenance' var for this worker. When True, this worker will not pick up any new requests. */
    maintenance?: boolean;
    /** The new state of the 'paused' var for this worker. When True, this worker will not be given any new requests. */
    paused?: boolean;
    /** The new state of the 'info' var for this worker. */
    info?: string;
    /** The new name for this this worker. */
    name?: string;
    /**
     * The new team of this worker.
     * @example Direct Action
     */
    team?: string;
  }
// ...
  export interface HordePerformance {
    /** The amount of waiting and processing image requests currently in this Horde */
    queued_requests?: number;
    /** The amount of waiting and processing text requests currently in this Horde */
    queued_text_requests?: number;
    /** How many workers are actively processing prompt generations in this Horde in the past 5 minutes */
    worker_count?: number;
    /** How many workers are actively processing prompt generations in this Horde in the past 5 minutes */
    text_worker_count?: number;
    /** How many worker threads are actively processing prompt generations in this Horde in the past 5 minutes */
    thread_count?: number;
    /** How many worker threads are actively processing prompt generations in this Horde in the past 5 minutes */
    text_thread_count?: number;
    /** The amount of megapixelsteps in waiting and processing requests currently in this Horde */
    queued_megapixelsteps?: number;
    /** How many megapixelsteps this Horde generated in the last minute */
    past_minute_megapixelsteps?: number;
    /** The amount of image interrogations waiting and processing currently in this Horde */
    queued_forms?: number;
    /** How many workers are actively processing image interrogations in this Horde in the past 5 minutes */
    interrogator_count?: number;
    /** How many worker threads are actively processing image interrogation in this Horde in the past 5 minutes */
    interrogator_thread_count?: number;
    /** The amount of tokens in waiting and processing requests currently in this Horde */
    queued_tokens?: number;
    /** How many tokens this Horde generated in the last minute */
    past_minute_tokens?: number;
  }
// ...
  export interface ImgModelStats {
    day?: SinglePeriodImgModelStats;
    month?: SinglePeriodImgModelStats;
    total?: SinglePeriodImgModelStats;
  }
  
  export interface SinglePeriodImgModelStats {
    "*"?: Record<string, number>;
  }